+++ docs/CHANGELOG.rst
@@ -14,6 +14,9 @@ Added
 -----
 - Expose ``status`` on ``collection`` and ``entity`` viewset and allow
   filtering and sorting by it
+Changed
+-------
+- Make processing container startup script ``Python`` 3.12 compatible
 
 
 ===================